Shader copying removed from build events (config file 1.2 now does the job)
git-svn-id: https://svn.vis.uni-stuttgart.de/projects/megamol/plugins/protein/trunk@16 5e8c2511-817d-0410-85a1-fd227d6480d1

+++ SolPathDataSource.cpp
@@ -97,7 +97,7 @@ bool SolPathDataSource::getData(megamol::core::Call &call) {
         this->loadData();
     }
 
-    spdc->Set(this->pathlines.Count(), this->pathlines.PeekElements(),
+    spdc->Set(static_cast<unsigned int>(this->pathlines.Count()), this->pathlines.PeekElements(),
         this->minTime, this->maxTime, this->minSpeed, this->maxSpeed);
 
     return true;
